fixed more OCD stuff. Moved API services into corresponding component etc
