+++ todo-app-standalone/src/components/modules/todo/api/todoService.js
@@ -1,4 +1,4 @@
-import ApiService from "./apiService"
+import ApiService from "../../../../api/services/apiService"
 const TODO_API_BASE_URL = "http://localhost:8080/api/todo"
 
 class TodoService extends ApiService {
