+++ todo-app-standalone/src/components/modules/user/api/UserApiService.js
@@ -1,4 +1,4 @@
-import ApiService from "./apiService"
+import ApiService from "../../../../api/services/apiService"
 
 const USER_API_BASE_URL = "http://localhost:8080/api/auth"
